Several improvements * Fix numeric precision issue * Fix error occurred when column name is sensitive

+++ src/model/user.ts
@@ -25,7 +25,7 @@ User.init({
         primaryKey: true,
     },
     cashBalance: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.DECIMAL,
         allowNull: false
     },
     name: {
